% Generated by build script
\name{s2_authors}
\alias{s2_authors}
\title{Search Semantic Scholar authors}
\usage{
s2_authors()
}
\arguments{
\item{query}{Author name to search}
\item{limit}{Maximum results (default 10, max 1000)}
}
\value{tibble: authorId, name, paperCount, citationCount, url}
\description{Search Semantic Scholar authors}
